+++ tests/lib/Integration/GeneratorsTest.php
@@ -426,9 +426,9 @@ class GeneratorsTest extends TestCase
     {
         $content = $this->assertAdapter();
 
-        $this->assertContains('Eloquent\AbstractAdapter', $content);
-        $this->assertNotContains('use DummyApp\Company;', $content);
-        $this->assertContains('parent::__construct(new \DummyApp\Company(), $paging);', $content);
+        $this->assertContentContains('Eloquent\AbstractAdapter', $content);
+        $this->assertContentNotContains('use DummyApp\Company;', $content);
+        $this->assertContentContains('parent::__construct(new \DummyApp\Company(), $paging);', $content);
     }
 
     /**
@@ -437,7 +437,7 @@ class GeneratorsTest extends TestCase
     private function assertGenericAdapter()
     {
         $content = $this->assertAdapter();
-        $this->assertContains('Adapter\AbstractResourceAdapter', $content);
+        $this->assertContentContains('Adapter\AbstractResourceAdapter', $content);
     }
 
     /**
@@ -457,12 +457,12 @@ class GeneratorsTest extends TestCase
         $content = $this->files->get($file);
 
         if ($this->byResource) {
-            $this->assertContains('namespace DummyApp\JsonApi\Companies;', $content);
-            $this->assertContains('class Adapter extends', $content);
+            $this->assertContentContains('namespace DummyApp\JsonApi\Companies;', $content);
+            $this->assertContentContains('class Adapter extends', $content);
         } else {
             $class = $this->withoutType ? 'Company' : 'CompanyAdapter';
-            $this->assertContains('namespace DummyApp\JsonApi\Adapters;', $content);
-            $this->assertContains("class {$class} extends", $content);
+            $this->assertContentContains('namespace DummyApp\JsonApi\Adapters;', $content);
+            $this->assertContentContains("class {$class} extends", $content);
         }
 
         return $content;
@@ -474,7 +474,7 @@ class GeneratorsTest extends TestCase
     private function assertEloquentSchema()
     {
         $content = $this->assertSchema();
-        $this->assertContains('return (string) $resource->getRouteKey();', $content);
+        $this->assertContentContains('return (string) $resource->getRouteKey();', $content);
     }
 
     /**
@@ -483,7 +483,7 @@ class GeneratorsTest extends TestCase
     private function assertGenericSchema()
     {
         $content = $this->assertSchema();
-        $this->assertNotContains('return (string) $resource->getRouteKey();', $content);
+        $this->assertContentNotContains('return (string) $resource->getRouteKey();', $content);
     }
 
     /**
@@ -502,16 +502,16 @@ class GeneratorsTest extends TestCase
         $this->assertFileExists($file);
         $content = $this->files->get($file);
 
-        $this->assertContains('extends SchemaProvider', $content);
-        $this->assertContains("protected \$resourceType = 'companies';", $content);
+        $this->assertContentContains('extends SchemaProvider', $content);
+        $this->assertContentContains("protected \$resourceType = 'companies';", $content);
 
         if ($this->byResource) {
-            $this->assertContains('namespace DummyApp\JsonApi\Companies;', $content);
-            $this->assertContains('class Schema extends', $content);
+            $this->assertContentContains('namespace DummyApp\JsonApi\Companies;', $content);
+            $this->assertContentContains('class Schema extends', $content);
         } else {
             $class = $this->withoutType ? 'Company' : 'CompanySchema';
-            $this->assertContains('namespace DummyApp\JsonApi\Schemas;', $content);
-            $this->assertContains("class {$class} extends", $content);
+            $this->assertContentContains('namespace DummyApp\JsonApi\Schemas;', $content);
+            $this->assertContentContains("class {$class} extends", $content);
         }
 
         return $content;
@@ -533,16 +533,16 @@ class GeneratorsTest extends TestCase
         $this->assertFileExists($file);
         $content = $this->files->get($file);
 
-        $this->assertNotContains('use DummyApp\Company;', $content);
-        $this->assertContains('@param mixed|null $record', $content);
+        $this->assertContentNotContains('use DummyApp\Company;', $content);
+        $this->assertContentContains('@param mixed|null $record', $content);
 
         if ($this->byResource) {
-            $this->assertContains('namespace DummyApp\JsonApi\Companies;', $content);
-            $this->assertContains('class Validators extends', $content);
+            $this->assertContentContains('namespace DummyApp\JsonApi\Companies;', $content);
+            $this->assertContentContains('class Validators extends', $content);
         } else {
             $class = $this->withoutType ? 'Company' : 'CompanyValidator';
-            $this->assertContains('namespace DummyApp\JsonApi\Validators;', $content);
-            $this->assertContains("class {$class} extends", $content);
+            $this->assertContentContains('namespace DummyApp\JsonApi\Validators;', $content);
+            $this->assertContentContains("class {$class} extends", $content);
         }
     }
 
@@ -565,12 +565,12 @@ class GeneratorsTest extends TestCase
         $this->assertFileExists($file);
         $content = $this->files->get($file);
 
-        $this->assertContains("class {$class} extends AbstractAuthorizer", $content);
+        $this->assertContentContains("class {$class} extends AbstractAuthorizer", $content);
 
         if ($this->byResource) {
-            $this->assertContains('namespace DummyApp\JsonApi;', $content);
+            $this->assertContentContains('namespace DummyApp\JsonApi;', $content);
         } else {
-            $this->assertContains('namespace DummyApp\JsonApi\Authorizers;', $content);
+            $this->assertContentContains('namespace DummyApp\JsonApi\Authorizers;', $content);
         }
     }
 
@@ -592,12 +592,12 @@ class GeneratorsTest extends TestCase
         $content = $this->files->get($file);
 
         if ($this->byResource) {
-            $this->assertContains('namespace DummyApp\JsonApi\Companies;', $content);
-            $this->assertContains('class Authorizer extends AbstractAuthorizer', $content);
+            $this->assertContentContains('namespace DummyApp\JsonApi\Companies;', $content);
+            $this->assertContentContains('class Authorizer extends AbstractAuthorizer', $content);
         } else {
             $class = $this->withoutType ? 'Company' : 'CompanyAuthorizer';
-            $this->assertContains('namespace DummyApp\JsonApi\Authorizers;', $content);
-            $this->assertContains("class {$class} extends", $content);
+            $this->assertContentContains('namespace DummyApp\JsonApi\Authorizers;', $content);
+            $this->assertContentContains("class {$class} extends", $content);
         }
     }
 
@@ -620,12 +620,12 @@ class GeneratorsTest extends TestCase
         $this->assertFileExists($file);
         $content = $this->files->get($file);
 
-        $this->assertContains("class {$class} extends BaseContentNegotiator", $content);
+        $this->assertContentContains("class {$class} extends BaseContentNegotiator", $content);
 
         if ($this->byResource) {
-            $this->assertContains('namespace DummyApp\JsonApi;', $content);
+            $this->assertContentContains('namespace DummyApp\JsonApi;', $content);
         } else {
-            $this->assertContains('namespace DummyApp\JsonApi\ContentNegotiators;', $content);
+            $this->assertContentContains('namespace DummyApp\JsonApi\ContentNegotiators;', $content);
         }
     }
 
@@ -647,12 +647,40 @@ class GeneratorsTest extends TestCase
         $content = $this->files->get($file);
 
         if ($this->byResource) {
-            $this->assertContains('namespace DummyApp\JsonApi\Companies;', $content);
-            $this->assertContains('class ContentNegotiator extends BaseContentNegotiator', $content);
+            $this->assertContentContains('namespace DummyApp\JsonApi\Companies;', $content);
+            $this->assertContentContains('class ContentNegotiator extends BaseContentNegotiator', $content);
         } else {
             $class = $this->withoutType ? 'Company' : 'CompanyContentNegotiator';
-            $this->assertContains('namespace DummyApp\JsonApi\ContentNegotiators;', $content);
-            $this->assertContains("class {$class} extends BaseContentNegotiator", $content);
+            $this->assertContentContains('namespace DummyApp\JsonApi\ContentNegotiators;', $content);
+            $this->assertContentContains("class {$class} extends BaseContentNegotiator", $content);
+        }
+    }
+
+    /**
+     * @param string $expected
+     * @param string $content
+     * @param string $message
+     */
+    private function assertContentContains($expected, $content, $message = '')
+    {
+        if (method_exists($this, 'assertStringContainsString')) {
+            $this->assertStringContainsString($expected, $content, $message);
+        } else {
+            $this->assertContains($expected, $content, $message);
+        }
+    }
+
+    /**
+     * @param string $expected
+     * @param string $content
+     * @param string $message
+     */
+    private function assertContentNotContains($expected, $content, $message = '')
+    {
+        if (method_exists($this, 'assertStringNotContainsString')) {
+            $this->assertStringNotContainsString($expected, $content, $message);
+        } else {
+            $this->assertNotContains($expected, $content, $message);
         }
     }
 }