+++ tests/lib/Unit/Store/StoreTest.php
@@ -50,7 +50,7 @@ class StoreTest extends TestCase
     /**
      * @return void
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->container = $this->createMock(ContainerInterface::class);