+++ tests/lib/Unit/Validators/ResourceDocumentValidatorTest.php
@@ -40,7 +40,7 @@ class ResourceDocumentValidatorTest extends TestCase
     /**
      * @return void
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->resourceTypes = ['people', 'tags', 'users'];